generate Cython callbacks
Summary:
This is the most complicated and magical code in the project. The C++ server is
calling a handler class that has a Py_Object attached to it. The C++ handler
class constructs a promise and passes it into the correct call_cy_* function
defined here. It also passes in the attached Py_object with the canonical name
`self` and any arguments that the service received.

At this point, all the arguments are still C++ objects, and we are running
outside the GIL. But now it's time to manipulate some Python code. So we grab
the gil and construct a lightweight Python wrapper around the promise so it can
be passed into a Python coroutine.

That coroutine is defined in this file as well and accepts the same self
parameter, the promise-wrapping Python object, and Python wrappers for any
arguments (these wrappers will be defined in later diffs; for now the code only
works when executed with no arguments).

The coroutine is executed on the asyncio event loop, which was stored as an
attribute on the Python handler object. Because the C++ server is not running on
the Python thread, we have to run it as a partial in a call to
call_soon_threadsafe (there is a new run_coroutine_threadsafe function included
with Python 3.5 that will do this more safely.)

Now that coroutine has a fairly simple job: First it awaits the user-defined
Python handler which does all the work and eventually returns a value. That
value is a Python object, so the coroutine has to convert it to the correct C++
object (this is not entirely implemented yet but it works for integers and
probably other primitives, but not void). The converted value is set on the
stored promise, which tells C++ to return the value asynchronously in the normal
way.

It's harder than it looks.

Reviewed By: cjtenny

Differential Revision: D3884051

fbshipit-source-id: 8694bab87cb5f6188d9f5fce0b7885eb276a2436