Hook up pyx file that is called from the user's Python code
Summary:
This diff brings together all the previous work. Essentially, it includes the
various pxi files and provides a handler that the Python user can extend and
implement.

This code is dependent on some "thriftlib" imports that will be supplied in a
later diff. Those paths will need to be changed once the dependent files have
been committed to an appropriate place in the thrift repository and we have
rearranged our build infrastructure so they can be found correctly.

Currently I have to include pxi files to get everything working correctly
because of some issues processing dependencies. I hope to move these to pxd's
that can be imported from in the near future.

Reviewed By: cjtenny

Differential Revision: D3884074

fbshipit-source-id: 68ab5bc190d7417487aba5001b5f92942f0be83d