Make getPendingCount work again (and turn it off by default)
Summary: A previous revision of mine (killing isUnevenLoad) broke this functionality. I'm not of the opinion this almost ever makes sense (for servers using lots of async_eb methods, this is helpful, for anyone else it's likely not). It consumes lots of cpu, even in its broken state, so gate it off. However, as it might be useful for some, make it so that it can be enabled by a flag.

Reviewed By: haijunz

Differential Revision: D3896949

fbshipit-source-id: 0dccda2cd0608053d3a908241905f1b0f67eb35a