Avoid a circular reference
Summary:
We need to use a weak reference to hold the Thrift server while it is starting up, to avoid a circular reference. Once the `ScopedServerThread::Helper` is unregistered as an event handler from the Thrift server, then we can (and should) go back to a owning smart pointer.

See https://our.intern.facebook.com/intern/sevmanager/view/s/140089/.

I originally tried to land this as D3823512, but that contained a use-after-free bug. This is the new implementation. Now the `EventHandler` class hands off control before it gets freed.

Reviewed By: haijunz

Differential Revision: D3838973

fbshipit-source-id: 286f71c1d151b4eec5f8df9e8a501a35416da833