Refactor legacy-reflection to compress error lines
Summary:
[Thrift] Refactor legacy-reflection to compress error lines.

When compiling programs that use legacy-reflection via static-reflection, it is ideal to reduce the number of compiler errors when there are types in play for which the static-reflection metatypes are missing. This diff does that - not 100%, but decently enough.

This helps immensely when the Thrift structures in play have many transitively-nested types from many transitively-included `.thrift` files to reduce the number of compiler-errors, helping the programmer find the signal in the noise of which are the types for which static-reflection metatypes are missing and must be added.

Reviewed By: juchem

Differential Revision: D3881977

fbshipit-source-id: 3d079102a6283215a8d5691a2856925266e8c1cb