Correctly generate structs assuming that shared_ptr references are used everywhere
Summary:
In order to support zero copy semantics, C++ references need to be shared_ptr
everywhere. This allows us to construct wrappers around structs that can be
owned independently in both Python/Cython land and C++ client/server land.We
have an in-progress patch that allows us to compile thrift files with shared_ptr
in all locations. This diff adjusts the cython wrappers to wrap the shared_ptr
reference instead of the original semantics.

Reviewed By: chadparry

Differential Revision: D3902370

fbshipit-source-id: d05a998b7983af42ca60d87eedf09c2e1322ca6a