Catch bugs in cpp2 even if mstch_cpp2 is working
Summary:
This fixes a bug in our tests that was introduced with mstch_cpp2. The tests will check the diff of the files after all code generator commands have completed, but by then the mstch_cpp2 generator overrides the value in the gen-cpp2 dir.

This moves mstch_cpp2 to generate in gen-mstch_cpp2 instead of gen-cpp2, and changes the test code to appropriately check values with gen-cpp2.

Reviewed By: chadparry

Differential Revision: D3936310

fbshipit-source-id: 2338d3b227c4b08e9ff77aca36bf8af501fdd29a