Adding active connection count to ThriftServer.
Summary:
This commit adds a new field to the Thrift observer (TServerObserver), activeConnections. It adds implementation in BaseThriftSever.h and BaseThriftServer.cpp.

The idea is to track the number of active connections as the Thrift server is running, much like we want to track active requests.

Note that it's an estimate of active connections, because we use a plain uint_32, not an std::atomic, so there may be occasional slight discrepancies in the count.

Also note this is my first time ever working with Thrift, and first time working with C++ since I was an undergrad ~20 years ago, so my syntax might be off. (That said, this code is pretty simple...)

Reviewed By: djwatson

Differential Revision: D3872659

fbshipit-source-id: d4355e40046263596ffecbd3c9acbf142052dc59