Add support for more primitive types
Summary:
This diff really highlights the genius of the templating design. While there are
a lot of templates to update, it is easy to see the symmetry with existing
types.

I've added support for bool, byte, and double types in addition to the three int
types that were already supported.

Reviewed By: yfeldblum

Differential Revision: D3912160

fbshipit-source-id: 7fe2ab0ba55dfc720e78d8e832c42a13688504ab