Flat view of a struct's getter hierarchy
Summary:
This diff implements a helper to retrieve a flat list of getters for the whole hierarchy of a Thrift structure.
With a single `get` it's possible to retrieve a deeply nested field.
This is useful when performing conversions from one struct hierarchy to another, or to perform format conversion on structs <-> maps.

Reviewed By: yfeldblum

Differential Revision: D3861852

fbshipit-source-id: ed5dcdd1a712160fbd2e525c25153682cd34d813