Adding flag `reflection`, soon to replace `fatal`
Summary:
This diff adds support for a new flag `reflection`, that will soon replace the flag `fatal`.
Fatal is just the library used to support static reflection, but that's incidental and could potentially change in the future. The long term feature is actually static reflection, and doesn't necessarily depend on Fatal to exist.

Reviewed By: yfeldblum

Differential Revision: D3866005

fbshipit-source-id: 90cb82f714f7d538f7365a0c94c284911437bfbf