thrift: t_cpp_generator.py: emit code with fewer shadowing warnings
Summary:
Fix a warning exposed by gcc-4.9.x's -Wshadow-compatible-local.
There were three "args" variables. I opted to rename the inner two (lambda
parameters) rather than the outer one, because the scope of each parameter
is narrower: guaranteed to extend only to the following line. In addition,
in the generating python, the code for the lambda parameters is all in
the same place, while the code that emits the declaration of the outer
"args" is far from the code that generates its use.

Likewise for the "prot" parameter.

Reviewed By: yfeldblum

Differential Revision: D3988253

fbshipit-source-id: cf90c9f47110dbef9d197e5a6077872903d506c1