Handle exceptions from handler
Summary:
When the python service handler code raises an exception, that exception needs
to be handled and returned to the client correctly using the existing C++
machinery. Currently it segfaults instead. This diff introduces handling that
catches any unexpected exceptions and sets the exception on the future
appropriately.

In future diffs, I will add code for handling expected exceptions (ie: those
that are predefined in the thrift interface file).

Reviewed By: yfeldblum

Differential Revision: D3915490

fbshipit-source-id: 3fa27d9a56f916803e0aeec47f52d0804f5e64d9