Support empty structs
Summary:
When I started working with exceptions, I discovered that structs with no
members are not being generated correctly. Luckily, there is a fixture for that!

In addition to supporting empty structs a little more cleanly, I made the
initializers put each argument on a second line, for consistency with other
generated calls.

Reviewed By: chadparry

Differential Revision: D3938283

fbshipit-source-id: f31bf728ee37128e2af23bb229b14095d3424fa6