Move ttypes definition from include to python extension
Summary:
The gist of this diff is to move the pxi file, which is included by Cython into
the main server (just like a C or mustache include) into into a separate pyx
file that can be imported by the server. This allows the types module to be
reused by the generated thrift-py3 client or directly from someone's Python code
without having to build the server.

Reviewed By: yfeldblum

Differential Revision: D3949294

fbshipit-source-id: 0105e61c36b01d31aab908cb3a6808f4928aeda8