move cython include statements to mustache-time includes
Summary:
This diff essentially uses mustache instead of cython to include content into
one file. This yields a double benefit: first, there can be more control of
namespacing in the future. Second, there are fewer files for the build system or
developer to manage when trying to cython-compile the generated output.

Reviewed By: yfeldblum

Differential Revision: D3952742

fbshipit-source-id: 0f7c76f636c72ab8871afe18ad3162bae07060d0