Add exception handling for thrift defined exceptions
Summary:
Now with more exception handling. This handles exceptions that the user expects
the service to throw, as defined in the thrift interface definition. I had to
make changes in quite a few places to make this work, but hopefully it won't be
too bad to review.

The implementation here catches each exception individually. I need to
restructure my structs a bit so that I can merge those into a library. That will
happen in a later diff because this one is hairy enough.

Reviewed By: yfeldblum

Differential Revision: D3955681

fbshipit-source-id: 8b7d86ad1f91d0e9d2a7f2773fe652edbca4f586