Collect all container types in one place
Summary:
I'm a bit sad at how complicated containers are going to have to be. The general
problem is that while Cython can easily wrap templated CPP classes and methods,
it is not able to define them in a generic way. That means that I need to
generate a new python extension class for every container type that we
encounter. Given that containers can contain other containers, you can see that
this gets hairy quickly.

After consulting with Haijun and Fried, we decided that the least bad solution
is to generate the containers in the modules that request them. This will make
included files tricky (they don't work correctly in this diff, but, then, none
of my code to date processes included files correctly).

Our plan is for the end developer to not have to know about this. Instead, we
will attempt to create a factory that can construct the correct wrapper class
when the user requests one. Because of duck typing, the user has no need to know
what kind of class cython is wrapping unless they are constructing it
themselves.

So this diff iterates over all the containers defined in an interface file,
whether used in args, return type, or struct references and creates a new
top-level unique set of containers that get generated along side the other types
for that if file.

Reviewed By: chadparry, haijunz

Differential Revision: D3995741

fbshipit-source-id: aa4b5bd30a8e90dd7a8cdc8a3bbc9cd8cb8b813f