Add connectionNewRequest to TServerEventHandler
Summary:
Add a new `connectionNewRequest` method in `TServerEventHandler` that will be called when a new `Cpp2Connection` starts working on a new `Cpp2Request`. This will help us better capture the lifecycle of a Thrift request for tracing and profiling. It takes parameter of a `TConnectionContext` of the `Cpp2Connection` and a `folly::RequestContext` of the request.

This Diff also changes the `folly::RequestContextScopeGuard` to use a explicitly created `folly::RequestContext`, the same one passed into the `connectionNewRequest` handler call. This would give us more ability to capture the setting of `folly::RequestContext` (the handler call happens before the first `folly:RequestContext::setContext` call invoked by the creation of `folly::RequestContextGuard`), especially for BPF profiling.

This Diff also moves the `folly::RequestContextScopeGuard` to an earlier location of the `Cpp2Connection::requestReceived` function so that it would better capture the lifecycle of an entire request.

Reviewed By: haijunz, drarmstr

Differential Revision: D3977022

fbshipit-source-id: 838a1767d498b6339a88da13ff437fffe76968ef