Add templates to support set and map structures
Summary:
Sets and maps are similar to lists, but this wasn't straightforward. For maps,
some of the included files that previously had hard-coded variable names had to
have custom names in the templates, one for keys and one for values. I had to
add a small change to the py3 generator that essentially allows the template to
have basic logic to flag whether or not a variable should be set at a specific
time. If that variable is set, it outputs "key", otherwise it outputs "item".
This isn't great, but it's the best solution I could come up with given the "no
logic" constraints of mustache.

I gave the new variable a ! suffix to indicate that it is sort of performing an
action on the current context. It has some symmetry to the existing ? suffix.

Other than that, the changes are parallel to what was needed for lists/vectors.

Reviewed By: yfeldblum

Differential Revision: D4015420

fbshipit-source-id: 08f5053f9eded8e53222d56aa1f8fc2d5973fb1e