Support child structs and strings
Summary:
A fairly simple change to allow struct attributes to refer to other structs or
to strings.

We had previously been planning to support "zero copy" by using shared_ptr in a
lot of places. However, this turns out to have a lot of implications for the
existing C++ code. In the future it will be useful to support thrift annotations
that use shared_ptr for references on a case-by-case basis as defined in the
thrift definition.

Reviewed By: yfeldblum

Differential Revision: D4029932

fbshipit-source-id: 808ff769d89276572809a28a471e38ef92225265