Ability to query worker count by priority on PriorityThreadManager
Summary: When using the PriorityThreadManager, if we want to dynamically change the # of workers per priority, we could query how many workers are there and call addWorker or removeWorker accordingly. To do that, we could expose a workerCount(PRIORITY) method instead of having the user keeping that count on a different variable.

Reviewed By: yfeldblum, juchem

Differential Revision: D4068848

fbshipit-source-id: e550607893ff3f489f5b0e166390d625646485e2