Use move semantics instead of shared_ptr on promise
Summary:
yfeldblum showed me how to pass a promise into a method using move semantics
while holding onto a reference to a future to return to the caller. I wasn't
sure if this would work with Cython as it doesn't, as a rule, support rvalue
references, but as with shared_ptr and unique_ptr, it's possible to work around
that with explicit definitions of move for each type.

Reviewed By: yfeldblum

Differential Revision: D4054376

fbshipit-source-id: 0e362e2f1b13a834cf3eed7dcc8200d45e0eb59d