Use the newly-constexpr enum_traits<T>::array::names in legacy_reflection
Summary:
[Thrift] Use the newly-`constexpr` `enum_traits<T>::array::names` in `legacy_reflection`.

Previously, it was not `constexpr`. This triggered a few problems when the enumeration being reflected was very large, so {D4079642} (5af16dcaaa974a773cfcc2a31f187f91caaf2018) hacked in a `constexpr` version of the `names` with both the strings and the sizes.

But, as of https://github.com/facebook/fatal/commit/478ef4dd42f8f3f7233b01fa5a714642a7e31da4, `enum_traits<T>::array::names` is `constexpr`, and it includes both the strings and the sizes. So we can use that directly.

Build times over large enumerations improve. Best guess is that a single param-pack expansion to compute a `constexpr string_view[]` performs better than two param-pack expansions, one to compute a `constexpr const char*[]` and one to compute a `constexpr uint16_t[]`, which the current code does.

Reviewed By: juchem

Differential Revision: D4094290

fbshipit-source-id: 056dc47a8fc977be96ebe83994ce8328c83e8f51