Return structs and containers
Summary:
So far the generated code has only been able to handle primitive types as return
values. This diff adds support for returning both structs and containers.

Unfortunately, because the generated cpp2 code expects a unique_ptr to the
return value being passed into the future, I have to do a copy of the value that
is passed in from Python. We will need to optimize this in order to better
support containers holding tens of thousands of large structs, for example.

In order to support structs and containers, I had to:

 *  make sure that promises returning containers have the appropriate name
 *  make sure that return values are properly generated for the new types
 *  convert an incoming container of items into the one to be returned
 *  force casts to check types, since we have no guarantees the user knows what
    they are doing

Reviewed By: haijunz

Differential Revision: D4059963

fbshipit-source-id: e2fa3f97c291b5a1689cf4b3e07abb5df2ccf3dd