Leak the non-constexpr generated constants
Summary:
[Thrift] Leak the non-constexpr generated constants.

Not much point in running all of their destructors at process exit - the whole address space is about to go away.

Options:
* The Leaky Meyers Singleton pattern: `static const T* const instance = new T(/*...*/); return *instance;`.
* The Indestructible Meyers Singleton pattern: `static const folly::Indestructible<T> instance(/*...*/); return *instance;`.

The effect is the same, but the Indestructible Meyers Singleton's object storage is inline, although the object may itself manage heap-allocated storage. The Leaky Meyers Singleton only stores a pointer inline, and stores the object on the heap. So the Indestructible Meyers Singleton pattern does one fewer heap-allocation per non-constexpr generated constant. It is unlikely to amount to very much of a difference compared to the Leaky Meyers Singleton pattern, but it might not be completely nothing.

(These are not the official names of any patterns. Just a bit of helpful nomenclature for the purpose of this discussion.)

Reviewed By: luciang

Differential Revision: D4103674

fbshipit-source-id: 0e23b884518c25653bc92455383f4dc1d10ba3ea