add support for enums
Summary:
Enum support requires a surprising number of changes, as we need to map between
Python enum types and c++ enum class types. Unfortunately, there doesn't seem to
be a way to convert between them using the intermediate integer type, so I've
added a helper function for each enum that converts between them.

You might ask why the helper function isn't a static method on the class: The
python Enum class is a pure-Python class (with some interesting metaclassing).
Extending a Python class in cython in such a way that a C++ method can be
accessed from c++ code is impossible.

Reviewed By: yfeldblum

Differential Revision: D4111412

fbshipit-source-id: e5241c679baf15f3bd41065f417c07ccc87db756