Unify cpp constant generation of enums
Summary:
[Thrift] Unify cpp constant generation of enums.

Do not crash the compiler with `enum_strict` - integral values casted to the type of a scoped enumeration, but which are not equal to any of the listed enumerators of the scoped enumeration, are nevertheless valid in C++.

Although this may be an invalid construction in other target languages. We may need to revisit this in the future.

Reviewed By: luciang

Differential Revision: D4116294

fbshipit-source-id: 23f9b49a3c2c7eb331c7030f074d178902456428