Support including structs from other thrift modules
Summary:
This is a bit of a messy diff, but it was easier than I expected.
The main difference is to ensure that an appropriate namespace is attached to
each struct and include that in the template instead of defaulting to the module
for the program that is being rendered. The appropriate namespace is already
available in thrift, so it just needed to be appended in extend_thrift. After
that, the only major change was to make sure that the correct imports were being
made at the top of the various modules.

In an upcoming diff I will add functionality to import and extend services as
well.

Reviewed By: cjtenny

Differential Revision: D4156411

fbshipit-source-id: 885c314497159827a8d8a75e4799341e4c95036d