Add support for nested containers
Summary:
Up until now, you could only use containers (lists, sets, and maps) that
contained primitive values and structs. This adds proper recursive template
inclusion support to handle nested containers correctly.

I'm doing one bad thing here; in order to properly recursively generate the
containers in the cython constructor, I create a new wrapper of the subcontainer
and then copy the data into a new vector. This is "necessary" because the user
can generate the containers from any Python iterable, and I need to make sure it
is backed by the proper c++ storage. I think there are ways to avoid the copies,
but they may require modifying the underlying c++ representation; we are
deferring this optimization for now.

This diff is not quite complete support for nested containers; I still need to
make sure the accessors are returning the correct values (ie: x[0] would return
a correctly wrapped list if x was a list of lists of something).

Reviewed By: yfeldblum

Differential Revision: D4195960

fbshipit-source-id: ad8a0024e8479fe53fc9ea905e1c08db24090035