Fix includes for missing namespaces
Summary:
Fix those fixtures that did not have a namespace defined. The compiler is
configured to set a default namespace if one is not supplied, much like the cpp2
compiler. When I added that fix, I neglected to change all the places that
touched the namespace.

I chose to make substantial changes to the templates to make this work instead
of having fragile hard-coded strings in the generator. Now, if we need to change
anything related to namespacing, it should only have to happen in one place
instead of these scattered locations.

A classic example of why DRY is important.

Reviewed By: yfeldblum

Differential Revision: D4202819

fbshipit-source-id: 8c8e85889a8d8262cd83dad46f7a531909361523