Enable hashing of structs
Summary:
This diff gives struct objects equality comparison by proxying through to the C
object's == operator. It also makes the struct hashable. This can be done
because the struct is immutable. Python 3 exceptions must be hashable (this is
one of the major broken features in the current thrift python implementation).
Structs aren't specifically required to be hashable, but there is no harm in
doing so, and hashing them has the added advantage of allowing us to memoize
functions that accept thrift objects, possibly yielding huge performance
savings.

For this iteration, I have implemented the hash function by proxying through to
Python's tuple hash function, which combines the hashes of the objects inside
the object. This is a common approach to hashing Python objects, but it is
suboptimal for this use-case. It would be much better to do the entire hash
calculation in C++ without.

Reviewed By: yfeldblum

Differential Revision: D4208556

fbshipit-source-id: c480aca951f829331312efc1cca9e014b881ebd4