Move main.h and generator functions to common
Summary:
* Moved externs from main.h to globals.h
* Deleted unnecessary declarations of common functions in main.h
* Moved validate_throws() from main to common

Note: Both the lex and the generator depend on this validate_throws(). This function had a dependency to a static function in generator called "get_true_type". Thus, I:
 * Copied the static get_true_type() function from generator to common and renamed common_get_true_type().
For now, I will keep the static get_true_type in generator for the sake of not having a gigantic diff. Next diff should remove this (and fix the broken dependencies).
* All of this gets rid of main.h
* Removed main.h from Makefile and includes.
* Fixed broken dependencies from removing main.h

Reviewed By: yfeldblum

Differential Revision: D4229465

fbshipit-source-id: 4c9a6a40448213cfc93fab9f15cd3bb01652f282