Make sure Cython can handle c++ exceptions while serving
Summary:
We get segfaults when, for example trying to reconnect a second service to the
same port. This clause adds a handler to the code that catches the C++
exceptions and reraises them as Python exceptions.

That should, in theory fix the problem but it doesn't. But they should be there
anyway.

Reviewed By: cjtenny

Differential Revision: D4241939

fbshipit-source-id: c6358d4f1601cee475889dba8e6fcf6642b80d3f