Add additional helper methods to Sets
Summary:
This diff fixes the broken contract on the lt/le/gt/ge methods, which are
supposed to do subset and superset comparisons. It also adds support for the
operations and methods used to do intersection, union, difference, and symmetric
difference.

When doing these latter tasks, it converts the incoming set to an instance of the
same class and then compares in C code.

Reviewed By: yfeldblum

Differential Revision: D4217426

fbshipit-source-id: 5c639b096db503512e6750ab014b6c7088409930