share backing store when constructing container types
Summary:
Users are generally expected to use Python's built-in container types until the
objects get into the cython/thrift/cpp machinery. At this point, the elements
are copied one by one into the appropriately typed wrapper objects.

However, there are quite a few places where this expensive copying is not
necessary. If a user receives a typed container and then sends it somewhere
else, for example, the same backing object can be used. More commonly, if a
container is attached to a struct or is inside a different container, it's
wrapper will be retreived from the parent object and then manipulated or passed
back into thrift directly.

Reviewed By: yfeldblum

Differential Revision: D4227284

fbshipit-source-id: b34eebeb12192df70360fe82772de8ef51d67836