Localize the type loops in the static-reflection libraries
Summary:
[Thrift] Localize the type loops in the static-reflection libraries.

Type loops are done using `fatal::foreach`. It needs a callable. Previously, the callable was implemented as a separate member type, named `visitor`, in the containing type with a templated `operator()`, or just as a templated member function `operator()`. However, that put the loop body far away from the `fatal::foreach` call-site, requiring the reader to jump around between the looping code and the loop body.

However, we can instead use a lambda expression with an `auto`-typed parameter and a by-reference capture default. This co-locates the loop body with the `fatal::foreach` call-site, and does not require the reader to jump around.

In exchange, we must pull type tagged type and the index value out of the parameter using `decltype`, instead of using type-level deconstrucing pattern-matching.

Before:

```lang=c++
struct container {
  using types = fatal::list</*...*/>; // the type-level container over which to iterate
  struct visitor {
    template <typename MyType, std::size_t Index, typename Arg>
    void operator()(fatal::indexed<MyType, Index>, Arg& arg) {
      // body of the runtime loop, using arg ...
    }
  };
  void go() {
    auto arg = result_of_many_lines_of_setup();
    fatal::foreach<types>(visitor{}, arg);
  }
};
```

After:

```lang=c++
struct container {
  using types = fatal::list</*...*/>; // the type-level container over which to iterate
  void go() {
    auto arg = result_of_many_lines_of_setup();
    fatal::foreach<types>([&](auto indexed) {
      using type = fatal::type_of<decltype(indexed)>;
      using index = decltype(indexed)::value; // std::size_t
      // body of the runtime loop, using arg ...
    });
  }
};
```

Reviewed By: juchem

Differential Revision: D4218873

fbshipit-source-id: 61b739df5fec24e7dd0669543e295e39324cbd88