thrift py3 asyncio: fix multiple known exceptions bug
Summary:
We're seeing clowny results when using a thrift asyncio service; a `JobAlreadyActive` exception every once in a while that does not have a `param` like `BadParameter` and dies on an internal layer.  It's ostensibly due to the `write_results_after_future` accepting the first known exception, not the appropriate one (which would pass a one-known-exception test).  Apparently it's random at runtime, which we suspect is due to dictionary ordering during thrift compilation & runtime jambalaya (`JobAlreadyActive` was first in code on my checkout, runtime `???`)

This does type checking beyond the except.

Reviewed By: scottruss

Differential Revision: D4255805

fbshipit-source-id: ae2e61241d5f4a69a00e4ed8ab9a330b31956878