Improve EOF error message.
Summary:
The EOF error message is a consistent source of questions within our internal user group. This change summarizes the three most common causes for EOF error messages on the client.

1) The server and the client have different idle timeouts.
2) The server has hit the connection limit, and is shedding idle connections.
3) The server crashed.

Although we have internal documentation on these issues, putting this short summary in the error message is likely to make things easier to debug for users.

Reviewed By: sriramsankar

Differential Revision: D4267862

fbshipit-source-id: 445081f0f6fb314db31a4c344517a710e0e71459