Compiler: Fixed broken Makefile
Summary:
Landing D4283051 uncovered a missing dependency in the Thrift Makefile base library

D4230226 added a folly dependency to the base library without adding the flag to the makefile. This was fine until I removed the Folly library that was stepping in later in the build process to fill in that dependency.

Reviewed By: chadparry

Differential Revision: D4296739

fbshipit-source-id: 822e8481a58286647bf0c4f2114fb384f42346d4