THeaderTransport.py Header Flags and Out of Order
Summary:
Expose access to flags, and provide helpers for setting and clearing flags

Header Sequence Numbering is important for Out of Order, so when
writing a message in THeaderProtocol.py we will copy our given seqid
to the Transport. This will not affect server which will parrot the
seqid it receives.

I set the out of order support flag in async_common, seems to be
best candidate for where to do it.

Reviewed By: haijunz

Differential Revision: D4303862

fbshipit-source-id: 0712706d7b3b98a56b28b35fa5c629be07a36155