THeaderTransport.py Remove a Buffer Copy
Summary:
Aliased 'readAll' to 'read', since the readAll in TTransportBase buffers
the data which THeaderTransport never needs to do unless its talking to a
depricated client or server. Instead we will do a readAll in our read for depricated
unframed. But everyone else will enjoy 1 less buffer copy when dealing with
THeaderTransport

Reviewed By: haijunz

Differential Revision: D4315836

fbshipit-source-id: 0e78eea3ce73c54083d12667e76a9528ea88cde9