Python 3 String and BytesIO
Summary:
THeaderTransport should decode info_header key and values.

And BytesStrIO should go away. If we ever try to shove a str into a BytesIO
in python 3 it is a bug in a thrift protocol or tansport and we need to fix it.
This just covers up those bugs at a lower level and slows down all IO in the process.

Reviewed By: haijunz

Differential Revision: D4316658

fbshipit-source-id: 6cd36002373c7d10cd704da2b8e2da8ad3bcb1d6