Compiler: Unit test the Validator
Summary:
This removed the usage of folly in validator.cc.
* folly/Format was replaced with std::string formatting
* folly/MapUtil::get_default was replaced with a ternary if to set a value.
* folly/Range::hasher was replaced with std::hash. Since we don't need to persist this data, std::hash is a good replacement as we are only using it to validate per environment.

A set of unit tests were added to test this behavior.
Note: A helper set of functions were added to allow the easy creation of fake objects to simply the test functions.

On the validator side, string formatting for error messages was moved to a helper function to ease readability in the main validator function.

Reviewed By: yfeldblum

Differential Revision: D4304224

fbshipit-source-id: 7cba1c5f25b62d16dc07f1ffa8993468cb817fd8