THeaderTransport.py support FRAMED_COMPACT and UNFRAMED_COMPACT
Summary:
Added support for FRAMED_COMPACT and UNFRAMED_COMPACT_DEPRECATED.
Also made sure we set the protocol id when we figure out what the
client type is, so THeaderProtocol can read from non-header clients
that have a protocol different than the default.

I added PROTOCOL_ID to TBinaryProtocol so the logic for detecting either
would be the same.

cleaned up readFrame so we only read from the transport what we need
to do during client_type detection, so we do less unpacking and less bitwise
manipulations.  The code should be easier to read.

Python3 bytes subscripts are ints so we make _ord = int, and since all
ints below 1024 are already created it should be mostly a no-op.
For Python2 subscripts of bytes are bytes so _ord = ord

Reviewed By: haijunz

Differential Revision: D4336456

fbshipit-source-id: a723a76d51a5e952cc5eb6769e7b088f7ef3baa6