Refactor some EventHandlerBaseTest setup code
Summary:
[Thrift] Refactor some `EventHandlerBaseTest` setup code.

And fix a bug in one case where an `exception_wrapper` is constructed from an `exception_ptr` and an `exception&`, but the `exception_ptr` and the `exception&` do not refer to the same object.

There is no harm //currently//. But it is incorrect, and a future change may begin enforcing the rule and cause the existing code to break.

Reviewed By: ericniebler

Differential Revision: D4363540

fbshipit-source-id: 0cd89c9b1966939dfe49adfeacf1f83415385a5a