Add cmake base to all fbthrift
Summary:
This adds a top level CMakeLists.txt file allowing to build from the top level directory using Cmake. I didn't change the top level Readme file yet as this only builds the compiler for now. I will slowly start adding CMakeLists.txt files on every required folder to build everything from the top level directory; or, build components of it(i.e. `cmake build_compiler ..`) from the top level directory as well.

Note: I had to change the executable name to `thriftcomp` as cmake makes a build tree inside build with the directories that have a `CMakeLists.txt` file. Having the binary named the same as de directory in the build tree would cause cmake to not create the binary.

There's 3 solutions for this naming problem:
1) Create the binary inside a `bin` directory in build `fbthrift/build/bin/trift`.
2) Keep the new name `thriftcomp` to have `fbthrift/build/thriftcomp`
3) Rename the directory to `fbthrift`. From `fbthrift/thrift` -> `fbthrift/fbthrift`.

I went for solution #2 but it's simple enough to move it to the other two.

Reviewed By: yfeldblum

Differential Revision: D4365856

fbshipit-source-id: 70481f0839ab28ed47fb0f9a28de119c54859038