Compiler: Fixed flex flags versioning issue
Summary:
Apparently if you use a flex version <2.6, the program can't handle empty strings if you specify that you will send a flag (i.e: `COMPILE_FLAGS`).

To fix I just moved the optional cmake argument `COMPILE FLAGS` to the if(WIN32) statement. In that way, if it's not windows, it will not expect any flags.

Reviewed By: yfeldblum

Differential Revision: D4379285

fbshipit-source-id: 382ae7c379a888cc02ab6e9aca226a3063220dd3