Implement two new container benchmarks for several servers
Summary:
Support a couple of benchmarks that are particularly interesting to the
thrift-py3 effort, specifically large containers. The load generator has been
updated in a separate diff. The goal of largeContainers to receive a large
structure and do nothing with it, which is basically the best case scenario for
thrift-py3. The iterAllFields method is the worst case scenario for thrift-py3,
in that all fields in all structs have to be converted to Python; the point of
interest is that it is still much faster than Asyncio in this case.

Reviewed By: yfeldblum

Differential Revision: D4329734

fbshipit-source-id: 5eb60c5640154d199131fb01c3e040e81c213c47