Compiler: Add proper const bool support
Summary:
The thrift compiler didn't have proper support for const values. They were treated as integers and a check was made if int == 0 to determine if true or false in code generation.

This changes that, this adds:
* A bool fields for t_const_value
* The changes for the scanner and parser to detect bool values
* The changes to mustache to use those boolean values.

There's a caveat: To maintain backward compatibility we must set the integer value in t_const_value. A new diff will be sent where we fix every code generator to check for the boolean rather than the integer.

Reviewed By: yfeldblum

Differential Revision: D4388843

fbshipit-source-id: 7383e8440149e98db6c2b578904950cb9193816b