Handle merge_into over ref-typed fields
Summary:
[Thrift] Handle `merge_into` over ref-typed fields.

Currently, `merge_into` fails to compile when instantiated for any type where any of its transitively nested fields is a ref-typed field. This diff gets it working.

Note that `merge` already is implemented for such cases ... in the naive way. So we write a set of test cases that replicate the existing implementation, and then write a parallel implementation for `merge_into`.

We also need to modify `pretty_print` to get the new test cases to compile, since the test cases use `pretty_print` to report errors.

Reviewed By: juchem

Differential Revision: D4391176

fbshipit-source-id: 8f5cb8e40747d0ad2b116969f2e22abc3410159d